--- a/app/models/document.rb
+++ b/app/models/document.rb
@@ -11,7 +11,7 @@
 #  size_bytes            :integer          not null
 #  created_at            :datetime         not null
 #  updated_at            :datetime         not null
-#  language_id           :bigint
+#  language_id           :bigint           not null
 #
 # Indexes
 #