--- a/db/migrate/20221201131258_set_column_language_on_document_to_not_null.rb
+++ b/db/migrate/20221201131258_set_column_language_on_document_to_not_null.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class SetColumnLanguageOnDocumentToNotNull < ActiveRecord::Migration[7.0]
+  def change
+    change_column_null :documents, :language_id, false
+  end
+end