--- a/spec/factories/languages.rb
+++ b/spec/factories/languages.rb
@@ -15,6 +15,5 @@
 #
 FactoryBot.define do
   factory :language do
-    iso_code { ISO_639.find('en').alpha3 }
   end
 end