--- a/spec/factories/languages.rb
+++ b/spec/factories/languages.rb
@@ -15,5 +15,6 @@
 #
 FactoryBot.define do
   factory :language do
+    iso_code { 'eng' }
   end
 end