--- a/spec/factories/pages.rb
+++ b/spec/factories/pages.rb
@@ -24,22 +24,34 @@ FactoryBot.define do
     document
     sequence(:page_num) { |n| n }
 
+    trait :image_generated do
+      after :create do |record, _evaluator|
+        filename = "#{File.basename(record.document.original_filename, '.pdf')}-#{record.page_num}.png"
+
+        record.image.attach(
+          io: file_fixture(filename).open,
+          filename:,
+        )
+        record.image_generated
+      end
+    end
+
     trait :text_extracted do
       text { 'Some mock text' }
 
       after :create do |record, _evaluator|
-        record.image_generated
         record.text_extracted
       end
     end
 
-    trait :image_generated do
+    trait :processed do
       after :create do |record, _evaluator|
-        record.image_generated
+        record.process
       end
     end
 
-    factory :page_with_text_extracted, traits: [:text_extracted]
+    factory :page_with_text_extracted, traits: %i[image_generated text_extracted]
     factory :page_with_image_generated_state, traits: [:image_generated]
+    factory :page_with_processed_state, traits: %i[image_generated text_extracted processed]
   end
 end