--- a/spec/factories/pages.rb
+++ b/spec/factories/pages.rb
@@ -35,19 +35,15 @@ FactoryBot.define do
       end
 
       after :create do |record, _evaluator|
-        puts 'hellllooooooooooooooooooooooooooo'
         record.image_generated
       end
     end
 
     trait :text_extracted do
       image_generated
-      # text { "This is the mock text for page #{page_num}" }
       after :create do |record, _evaluator|
         record.update(text: "This is the mock text for page #{record.page_num}")
         record.text_extracted
-        # record.reload
-        puts "************************** #{record.id}: #{record.current_state}"
       end
     end