--- a/spec/jobs/text_extraction_job_spec.rb
+++ b/spec/jobs/text_extraction_job_spec.rb
@@ -16,7 +16,6 @@ RSpec.describe TextExtractionJob do
           trust the people who wrote the software. KEN THOMPSON
         OCR
       end
-      let(:language) { create(:language) }
 
       context 'when there is text embedded' do
         let!(:document) do