… Gemfile.lock
… spec/fixtures/files/8000ad-1.png
… spec/fixtures/files/accented-1.png
… spec/fixtures/files/p761-thompson-1.png
… spec/fixtures/files/p761-thompson-2.png
… spec/fixtures/files/p761-thompson-3.png
… spec/fixtures/files/p761-thompson-notext-1.png
… spec/fixtures/files/p761-thompson-notext-2.png
… spec/fixtures/files/p761-thompson-notext-3.png
… spec/system/searches_spec.rb
… spec/time_formatter.rb