Close #28 

Before:
```
Finished in 15.48 seconds (files took 2.76 seconds to load)
```

Now:
```
Finished in 9.06 seconds (files took 2.64 seconds to load)
```